Define all service functions in array
This change started while looking at a CI failure when running
core_externallib_testcase related to a constant being defined more
than once. I was initially going to create a separate constants.php,
but then realized the approach of defining functions for the service
in two ways (the array list with built-in functions and the service
reference when declaring functions) was a bit odd. I think this
approach makes it easier to see / maintain what's going on and should
also fix the test issue.